RAG_567Task/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.5.0
plotly>=5.0.0
scikit-learn>=1.0.0
nltk>=3.8.0
tiktoken>=0.4.0
transformers>=4.30.0
torch>=1.13.0
matplotlib>=3.5.0
seaborn>=0.11.0
langchain>=0.0.300
langchain-text-splitters>=0.0.1
python-docx>=0.8.0
PyPDF2>=3.0.0
